Fix issue of shop=null on some NavigationMenu redirects (#37)
We've been getting reports of people using the new left hand navigation menu, and ending up with an empty screen in the app. The main app page loads fine, but any navigation using the menu causes this. We've tracked it down to the shop param sometimes being missing in `navigation.js`, and so they end up with shop=null. This commit adds a fallback method so we grab the myshopify.com from the dom if it's missing in the url params. Could be cleaner, so open to ideas.